Currenciesのモジュール上の位置を変更
