+++ src/main.rs
@@ -1,6 +1,6 @@
 extern crate zaif_api;
 
-use zaif_api::Currencies;
+use zaif_api::public_api::currencies::Currencies;
 
 fn main() {
     let api = Currencies{};
